requirements.txt
pandas
numpy
matplotlib
flask  
scikit-learn
Pillow